refactor(review): 変更ナラティブのコメント除去と ComfortableRange の非公開化
レビュー対応。Validate のコメントから変更経緯「query に降りた」を除き定常状態の説明にする。
ComfortableRange は query 内でしか使わず公開不要なので comfortableRange へ落とし、
直接の単体テストは内部テストへ移す。Insulation は公開関数の戻り型で外部が参照するため公開を維持する。
